+++ woodwork/tests/testing_utils/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 from .column_utils import convert_series
-from .datatable_utils import (
+from .table_utils import (
     is_property,
     is_public_method,
     mi_between_cols,
